Define standard parameters through individual mixins
