+++ docs/reference/index.rst
@@ -2,33 +2,40 @@
 Reference
 =========
 
-Base classes
-============
+Spiders
+=======
 
-.. autopydantic_model:: zyte_spider_templates.spiders.base.BaseSpiderParams
-    :inherited-members: BaseModel
+.. autoclass:: zyte_spider_templates.BaseSpider
 
-.. autoclass:: zyte_spider_templates.spiders.base.BaseSpider
+.. autoclass:: zyte_spider_templates.EcommerceSpider
 
-.. autoenum:: zyte_spider_templates.spiders.base.ExtractFrom
-    :noindex:
-
-.. autoenum:: zyte_spider_templates.spiders.base.Geolocation
-    :noindex:
-
-E-commerce
-==========
-
-.. autopydantic_model:: zyte_spider_templates.spiders.ecommerce.EcommerceSpiderParams
-    :noindex:
-    :inherited-members: BaseModel
-
-.. autoenum:: zyte_spider_templates.spiders.ecommerce.EcommerceCrawlStrategy
-    :noindex:
-
-.. autoclass:: zyte_spider_templates.spiders.ecommerce.EcommerceSpider
 
 Pages
 =====
 
 .. autoclass:: zyte_spider_templates.pages.HeuristicsProductNavigationPage
+
+
+Parameter mixins
+================
+
+.. autopydantic_model:: zyte_spider_templates.params.ExtractFromParam
+    :exclude-members: model_computed_fields
+
+.. autoenum:: zyte_spider_templates.params.ExtractFrom
+
+.. autopydantic_model:: zyte_spider_templates.params.GeolocationParam
+    :exclude-members: model_computed_fields
+
+.. autoenum:: zyte_spider_templates.params.Geolocation
+
+.. autopydantic_model:: zyte_spider_templates.params.MaxRequestsParam
+    :exclude-members: model_computed_fields
+
+.. autopydantic_model:: zyte_spider_templates.params.UrlParam
+    :exclude-members: model_computed_fields
+
+.. autopydantic_model:: zyte_spider_templates.spiders.ecommerce.EcommerceCrawlStrategyParam
+    :exclude-members: model_computed_fields
+
+.. autoenum:: zyte_spider_templates.spiders.ecommerce.EcommerceCrawlStrategy
