+++ docs/templates/e-commerce.rst
@@ -16,9 +16,4 @@ Parameters
 
 .. autopydantic_model:: zyte_spider_templates.spiders.ecommerce.EcommerceSpiderParams
     :inherited-members: BaseModel
-
-.. autoenum:: zyte_spider_templates.spiders.ecommerce.EcommerceCrawlStrategy
-
-.. autoenum:: zyte_spider_templates.spiders.base.ExtractFrom
-
-.. autoenum:: zyte_spider_templates.spiders.base.Geolocation
+    :exclude-members: model_computed_fields
