+++ zyte_spider_templates/__init__.py
@@ -1,2 +1,6 @@
 from .spiders.base import BaseSpider, BaseSpiderParams
-from .spiders.ecommerce import EcommerceSpider
+from .spiders.ecommerce import (
+    EcommerceCrawlStrategy,
+    EcommerceCrawlStrategyParam,
+    EcommerceSpider,
+)
